+++ permission_handler_android/lib/permission_handler_android.dart
@@ -1,0 +1,5 @@
+export 'src/android_object_mirrors/activity.dart';
+export 'src/android_object_mirrors/activity_compat.dart';
+
+export 'src/android.dart';
+export 'src/permission_handler_android.dart';